fix: handle cross-platform type differences in disk usage calculations
- Convert stat.Blocks and stat.Bavail to uint64 for consistent arithmetic
- Add unconvert ignore comments for platform-specific type differences
- Linux and macOS have different underlying types for syscall.Statfs_t fields
- This ensures compatibility across all supported Unix platforms

+++ internal/services/diskusage_unix.go
@@ -17,10 +17,10 @@ func getDiskUsageForPath(path string) (*DiskUsage, error) {
 	// Calculate sizes in bytes
 	//nolint:gosec // Safe conversion for disk space calculation
 	blockSize := uint64(stat.Bsize)
-	//nolint:gosec // Safe conversion for disk space calculation
-	totalSize := int64(stat.Blocks * blockSize)
-	//nolint:gosec // Safe conversion for disk space calculation
-	freeSize := int64(stat.Bavail * blockSize)
+	//nolint:gosec,unconvert // Safe conversion for disk space calculation, types vary by platform
+	totalSize := int64(uint64(stat.Blocks) * blockSize)
+	//nolint:gosec,unconvert // Safe conversion for disk space calculation, types vary by platform
+	freeSize := int64(uint64(stat.Bavail) * blockSize)
 
 	return &DiskUsage{
 		Free:  freeSize,
